requirements.txt
SQLAlchemy
sqlalchemy-utils
python-dotenv
pymysql
cryptography
requests